requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
python-telegram-bot[job-queue]>=21.0
tavily-python>=0.5.0
py-clob-client>=0.34.0
ccxt>=4.4.0
python-dotenv>=1.0.0
